requirements.txt
ultralytics>=8.3.0
supervision>=0.21.0
opencv-python>=4.10.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
scipy>=1.10.0
